research/parallax-effect/depth-map-generator/pyproject.toml
[project]
name = "depth-map-generator"
version = "0.1.0"
description = "Tool for generating depth maps from 2D images using Depth Anything V2 or MiDaS"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "Pillow>=9.0.0",
    "numpy>=1.24.0",
    "opencv-python>=4.7.0",
    "timm>=0.9.0",
]
